Improved graphic pane layout; Added metadata Some changes from `int` to `size_t`

+++ bruteforce.h
@@ -17,7 +17,7 @@ Solution bruteForceSearch(std::vector<Coordinate> &cities, int startCity,
                           std::function<bool()> stopRequested ) {
     Solution currentSol;
     currentSol.path.push_back(startCity);
-    for(int i = 0; i < cities.size(); i++)
+    for(size_t i = 0; i < cities.size(); i++)
         if(i != startCity)
             currentSol.path.push_back(i);
